Revert "Update message building and unit tests for Core project"
This reverts commit 7851725b5a37cdbc9b0d413c38cd3e11d3634468.
